JS-7623: space index header more
