+++ src/ts/component/header/main/settings.tsx
@@ -1,14 +1,17 @@
 import React, { forwardRef, useState } from 'react';
 import { observer } from 'mobx-react';
-import { I, S, U, translate, Relation, analytics } from 'Lib';
+import { I, S, U, translate, Relation, analytics, Action } from 'Lib';
 import { Icon, Label, Title, Button } from 'Component';
+import $ from 'jquery';
 
 const HeaderMainSettings = observer(forwardRef<{}, I.HeaderComponent>((props, ref) => {
-	const { renderLeftIcons } = props;
+	const { renderLeftIcons, menuOpen } = props;
 	const { account } = S.Auth;
 	const profile = U.Space.getProfile();
 	const participant = U.Space.getParticipant() || profile;
 	const globalName = Relation.getStringValue(participant?.globalName);
+	const space = U.Space.getSpaceview();
+	const isOwner = U.Space.isMyOwner(space.targetSpaceId);
 
 	const onIdentity = () => {
 		if (globalName) {
@@ -21,6 +24,30 @@ const HeaderMainSettings = observer(forwardRef<{}, I.HeaderComponent>((props, re
 		});
 
 		analytics.event('ClickUpgradePlanTooltip', { type: 'identity' });
+	};
+
+	const onMore = () => {
+		menuOpen('select', '#button-header-more', {
+			horizontal: I.MenuDirection.Right,
+			data: {
+				options: [
+					{ id: 'spaceInfo', name: translate('popupSettingsSpaceIndexSpaceInfoTitle') },
+					{ id: 'delete', name: isOwner ? translate('pageSettingsSpaceDeleteSpace') : translate('commonLeaveSpace'), color: 'red' },
+				],
+				onSelect: (e: React.MouseEvent, option: any) => {
+					switch (option.id) {
+						case 'spaceInfo': {
+							Action.spaceInfo();
+							break;
+						};
+						case 'delete': {
+							Action.removeSpace(S.Common.space, analytics.route.settings);
+							break;
+						};
+					};
+				},
+			}
+		});
 	};
 
 	const renderIdentity = () => {
@@ -40,11 +67,31 @@ const HeaderMainSettings = observer(forwardRef<{}, I.HeaderComponent>((props, re
 		);
 	};
 
+	const renderMore = () => {
+		const param = U.Router.getParam(U.Router.getRoute());
+		const id = param.id || 'account';
+		const showId = [ 'spaceIndex' ];
+
+		if (!showId.includes(id)) {
+			return '';
+		};
+
+		return (
+			<Icon
+				id="button-header-more"
+				tooltipParam={{ text: translate('commonMenu'), typeY: I.MenuDirection.Bottom }}
+				className="more withBackground"
+				onClick={onMore}
+				onDoubleClick={e => e.stopPropagation()}
+			/>
+		);
+	};
+
 	return (
 		<>
 			<div className="side left">{renderLeftIcons(true)}</div>
 			<div className="side center">{renderIdentity()}</div>
-			<div className="side right" />
+			<div className="side right">{renderMore()}</div>
 		</>
 	);
